removed some lib functions
