+++ package.yaml
@@ -23,6 +23,8 @@ dependencies:
   - base >= 4.7 && < 5
   - containers
   - vector
+  - extra
+  - safe
 
 library:
   source-dirs: src
